+++ R/read_delim.R
@@ -379,8 +379,8 @@ read_tokens <- function(data, tokenizer, col_specs, col_names, locale_, n_max, p
 }
 
 read_delimited <- function(file, tokenizer, col_names = TRUE, col_types = NULL,
-                           locale = default_locale(), skip = 0, skip_empty_rows = TRUE, comment = "",
-                           n_max = Inf, guess_max = min(1000, n_max), progress = show_progress()) {
+                           locale = default_locale(), skip = 0, skip_empty_rows = TRUE, skip_quote = TRUE,
+                           comment = "", n_max = Inf, guess_max = min(1000, n_max), progress = show_progress()) {
   name <- source_name(file)
   # If connection needed, read once.
   file <- standardise_path(file)
@@ -405,11 +405,12 @@ read_delimited <- function(file, tokenizer, col_names = TRUE, col_types = NULL,
   spec <- col_spec_standardise(
     data,
     skip = skip, skip_empty_rows = skip_empty_rows,
+    skip_quote = skip_quote,
     comment = comment, guess_max = guess_max, col_names = col_names,
     col_types = col_types, tokenizer = tokenizer, locale = locale
   )
 
-  ds <- datasource(data, skip = spec$skip, skip_empty_rows = skip_empty_rows, comment = comment)
+  ds <- datasource(data, skip = spec$skip, skip_empty_rows = skip_empty_rows, comment = comment, skip_quote = skip_quote)
 
   if (is.null(col_types) && !inherits(ds, "source_string") && !is_testing()) {
     show_cols_spec(spec)